theauth/models.py
from django.db import models
from django.conf import settings
from django.db.models.signals import post_save
from rest_framework.authtoken.models import Token
from django.dispatch import receiver

# Create your models here.
@receiver(post_save, sender=settings.AUTH_USER_MODEL)
def create_auth_token(request, instance=None, created=False, **kwargs):
    if created:
        Token.objects.create(user=instance)
