requirements.txt
dj-database-url==0.5.0
Django==1.11
django-heroku==0.3.1
djangorestframework==3.9.4
gunicorn==19.9.0
Pillow==6.0.0
psycopg2==2.8.2
python-decouple==3.1
pytz==2019.1
whitenoise==4.1.2
